+++ js/backend_upload.js
@@ -15,14 +15,14 @@
 	   {
          $.insert('{LEPTON_URL}/modules/modul_info/js/overlib.js');
 	   }
-       if($("#upload_wrap").length)
+       if($("#upload_wrap_{SECTION_ID}").length)
 	   {
 			$.insert('{LEPTON_URL}/modules/lib_jquery/jquery-ui/external/jquery.MultiFile.pack.js');
 			$.insert('{LEPTON_URL}/modules/lib_jquery/jquery-ui/external/jquery.MetaData.js');
-			$('#upload_wrap').removeClass('hidden');
+			$('#upload_wrap_{SECTION_ID}').removeClass('hidden');
 			$('.noscript').removeClass('hidden');
 			$('#file_script').MultiFile({
-				list: '#upload_wrap',
+				list: '#upload_wrap_{SECTION_ID}',
 			 STRING: {
 			  file: '<em title="Click to remove" onclick="$(this).parent().prev().click()">$file<\/em>',
 			  remove: '<img src="{THEME_URL}/images/delete_16.png" height="16" width="16" alt="x"/>'